Add arc2x2_sensLAr.gdml
This is the standalone 2x2 geometry (taken from GeoMergeFor2x2),
modified only to add a SensDet to volLArActive. We're not using this for
the main (2x2+MINERvA) simulation, but we are using it for the particle
bomb simulation, and potentially others in the future, so might as well
keep it in the `main` branch.