Update reaction code to match CAFs
Change the reaction code/variable in the MC truth dataset to match what
is used for CAFs. This numbering comes from the GENIE enum definition
for scattering modes.

For now the isXYZ reaction flags are still saved. May be removed in the
future.

GENIE scattering modes/types can be found here:
https://github.com/GENIE-MC/Generator/blob/master/src/Framework/Interaction/ScatteringType.h